Simplify by using latest features of mr.awsome and mr.awsome.ansible.
The ``bootstrap-jailhost`` command is not just called ``bootstrap``.
It always uses the ``bootstrap`` command now, the different bootstrap
types are selected by different default fabric files which import the
correct Fabric task as ``bootstrap``.

The ``configure-jailhost`` command is replaced by better options of the
``configure`` command of mr.awsome.ansible. It now sets the ``roles``
option if a playbook isn't available already.

All the hooks are replaced by a single ``augment_instance`` plugin
function which modifies the instance config as needed for all instances.