allow to optionally specify multiple NAT interfaces
the typical use-case for this is when you have a VirtualBox
instance with two interfaces - one host-only interface and one NATted
interface.

if only one of these interfaces are NATted you either won't be able
to access jails on the host via the host-only interface or the jails
won't be able to communicate with the internet.

only configuring both interfaces for NAT via ``pf`` allows for both
to take place.

to set more than one interface, simply define ``pf_nat_interfaces``
in a playbook.

to set only one interface (i.e. if you don't want the default) set
``pf_nat_interface`` (note the singular) as usual.